Limiting Slider only for Home page
